+++ components/ui/header.tsx
@@ -4,6 +4,7 @@ import UserImg from "@/public/images/user-image.jpg";
 import HeaderImg04 from "@/public/images/header-image-04.webp";
 import HeaderImg05 from "@/public/images/header-image-05.gif";
 import HeaderImg06 from "@/public/images/header-image-06.webp";
+import Link from "next/link";
 
 export default function Header() {
   return (
@@ -12,14 +13,16 @@ export default function Header() {
       <ThemeToggle />
       {/* Intro */}
       <div className="mb-10">
-        <Image
-          className="inline-flex rounded-full shadow-lg mb-4"
-          src={UserImg}
-          width={48}
-          height={48}
-          alt="Kyle Ross"
-          priority
-        />
+        <Link href="/">
+          <Image
+            className="inline-flex rounded-full shadow-lg transition duration-150 ease-in-out transform hover:scale-105 hover:shadow-md dark:hover:shadow-lg mb-4"
+            src={UserImg}
+            width={48}
+            height={48}
+            alt="Kyle Ross"
+            priority
+          />
+        </Link>
         <div className="mb-5">
           <h1 className="font-inter-tight font-bold text-gray-800 dark:text-gray-100 text-2xl mb-1">
             Kyle Ross
@@ -51,6 +54,7 @@ export default function Header() {
           height={160}
           alt="Header 02"
           priority
+          unoptimized
         />
         <Image
           className="rounded-xl even:rotate-2 odd:-rotate-2 group-hover:rotate-0 transition duration-300 ease-[cubic-bezier(.5,.85,.25,1.8)] shadow-lg"